Moved IGitFlow to Workflows namespace Added IGitFlowWithPullRequest Added IGitHubFlowWithPullRequest
