+++ src/Candoumbe.Pipelines/Candoumbe.Pipelines.csproj
@@ -8,6 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.DependencyValidation.Analyzers" Version="0.11.0" />
     <PackageReference Include="Nuke.Common" Version="6.2.1" />
   </ItemGroup>
 </Project>
