+++ src/Candoumbe.Pipelines/Components/IHaveGitVersion.cs
@@ -3,6 +3,8 @@
 using Nuke.Common;
 using Nuke.Common.Tools.GitVersion;
 
+using Octokit;
+
 namespace Candoumbe.Pipelines.Components;
 
 /// <summary>
@@ -19,7 +21,16 @@ public interface IHaveGitVersion : INukeBuild
     public GitVersion GitVersion => TryGetValue(() => GitVersion);
 
     /// <summary>
+    /// Hint to create a major release.
+    /// </summary>
+    /// <remarks>
+    /// The value of this property is only taken into account when running <see cref="Release"/> target.
+    /// </remarks>
+    [Parameter("Hint to create a major release.")]
+    bool Major => false;
+
+    /// <summary>
     /// Major.Minor.Patch Version number
     /// </summary>
-    string MajorMinorPatchVersion => GitVersion.MajorMinorPatch;
+    string MajorMinorPatchVersion => Major ? $"{GitVersion.Major + 1}.0.0" : GitVersion.MajorMinorPatch;
 }
