+++ src/Candoumbe.Pipelines/Components/Workflows/IHaveMainBranch.cs
@@ -1,0 +1,12 @@
+﻿namespace Candoumbe.Pipelines.Components.Workflows;
+
+/// <summary>
+/// Defines a single branch name for main
+/// </summary>
+public interface IHaveMainBranch
+{
+    /// <summary>
+    /// Name of the main branch
+    /// </summary>
+    const string MainBranchName = "main";
+}
